+++ materialsystem/cmatrendercontext.cpp
@@ -2200,7 +2200,7 @@ int CMatRenderContext::CompareMaterialCombos( IMaterial *pMaterial1, IMaterial *
 	if ( dLightmap )
 		return dLightmap;
 
-	return (int)pMat1 - (int)pMat2;
+	return (intp)pMat1 - (intp)pMat2;
 }
